products/requirements.txt
algoliasearch==2.6.3
algoliasearch-django==2.0.0
appdirs==1.4.4
asgiref==3.7.2
beautifulsoup4==4.12.3
black==23.12.1
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cssselect==1.2.0
Django==5.0.1
django-cors-headers==4.3.1
django-filter==23.5
django-rest-passwordreset==1.4.0
djangorestframework==3.14.0
fake-useragent==1.5.0
idna==3.6
importlib_metadata==7.0.2
lxml==5.1.0
Markdown==3.5.2
mypy-extensions==1.0.0
packaging==23.2
parse==1.20.1
pathspec==0.12.1
platformdirs==4.1.0
pyee==11.1.0
pyppeteer==2.0.0
pyquery==2.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2023.4
urllib3==1.26.18
w3lib==2.1.2
websockets==10.4
zipp==3.17.0
